+++ apps/cookbook/src/app/examples/checkbox-example/checkbox-example.component.ts
@@ -7,7 +7,7 @@ import { CheckboxStatesExampleComponent } from './examples/states';
 import { CheckboxSizesExampleComponent } from './examples/sizes';
 import { CheckboxMultilineExampleComponent } from './examples/multiline';
 import { CheckboxEventsExampleComponent } from './examples/events';
-import { CheckboxIndeterminateListExampleComponent } from '~/app/examples/checkbox-example/examples/list-indetermined';
+import { CheckboxIndeterminateListExampleComponent } from '~/app/examples/checkbox-example/examples/list-indeterminate';
 
 @Component({
   selector: 'cookbook-checkbox-example',
