+++ apps/cookbook/src/app/examples/checkbox-example/examples/list-indeterminate.ts
