+++ apps/cookbook/src/app/showcase/checkbox-showcase/checkbox-showcase.component.html
@@ -36,7 +36,7 @@
   <cookbook-checkbox-list-example #checkboxListExample></cookbook-checkbox-list-example>
 </cookbook-example-viewer>
 
-<h2>Indetermined state</h2>
+<h2>indeterminate state</h2>
 <p>
   In a list of checkboxes, the "Select all" checkbox can be in an indeterminate state. This means
   that neither all or none are selected.
